+++ export/orbax/export/obm_export.py
@@ -14,10 +14,13 @@
 
 """Export class that implements the save and load abstract class defined in Export Base for use with the Orbax Model export format."""
 
-from typing import Any
+from typing import Any, Union, cast
 
 from absl import logging
+from orbax.export import constants
 from orbax.export import export_base
+from orbax.export.modules import obm_module
+from orbax.export.modules import orbax_module_base
 import tensorflow as tf
 
 
@@ -26,25 +29,25 @@ class ObmExport(export_base.ExportBase):
 
   def save(
       self,
-      jax_module: tf.Module,
+      jax_module: Union[orbax_module_base.OrbaxModuleBase, tf.Module],
       model_path: str,
-      **kwargs: Any,
   ):
     """Saves a Jax model in the Orbax Model export format.
 
     Args:
       jax_module: The `JaxModule` to be exported.
       model_path: The path to save the model.
-      **kwargs: Additional arguments to pass to the `save` method. Accepted
-        arguments are `save_options` and `serving_signatures`.
+      save_options: The `SaveOptions` to use when exporting the model.
     """
 
-    # TODO: b/363061755 - Implment the ObmExport.save method.
-    logging.info("Exporting model using Orbax Export Model.")
-    raise NotImplementedError("ObmExport.save not implemented yet.")
+    if jax_module.export_version() != constants.ExportModelType.ORBAX_MODEL:
+      raise ValueError(
+          "JaxModule is not of type ORBAX_MODEL. Please use the correct"
+          " export_version. Expected ORBAX_MODEL, got"
+          f" {jax_module.export_version()}"
+      )
 
   def load(self, model_path: str, **kwargs: Any):
     """Loads the model previously saved in the Orbax Model export format."""
-    # TODO: b/363061755 - Implment the ObmExport.load method.
     logging.info("Loading model using Orbax Export Model.")
     raise NotImplementedError("ObmExport.load not implemented yet.")